added Shooter setSpeed func,# hopper indexer
added a function for the shooter subsystem to just take a speed and set
the shooterMotor speed to it.
Changed the initialization and number of hopper indexing servos.
Now they are initialized inside of an array that iterates from 0->3, and
assigns at 5 + i (that is, the first hopper is initialized to PWM 5.
Changed the hopper toggle to reflect this.